#3 refactoring user service
