+++ src/main/java/xyz/hlafaille/seizuretracker/exception/UserEntityMissingException.java
@@ -1,0 +1,7 @@
+package xyz.hlafaille.seizuretracker.exception;
+
+public class UserEntityMissingException extends Exception{
+    public UserEntityMissingException() {
+        super("User entity is missing");
+    }
+}
