+++ src/main/java/xyz/hlafaille/seizuretracker/service/AuthService.java
@@ -10,11 +10,6 @@ import java.util.UUID;
  * @version 1.0.0
  */
 public interface AuthService {
-    User getUserById(UUID id);
-
-    @Deprecated
-    UUID createUser(String firstName, String lastName, String email, String password);
-
     /**
      * Encrypt the given password using argon2
      *
@@ -30,12 +25,4 @@ public interface AuthService {
      */
     void matchPassword(User user, String password) throws RuntimeException;
 
-    /**
-     * Start a new session by logging in
-     *
-     * @param emailAddress User's email address
-     * @param password     Users unencrypted password
-     * @return Session UUID
-     */
-    UUID beginSession(String emailAddress, String password);
 }
