+++ src/main/java/xyz/hlafaille/seizuretracker/service/UserService.java
@@ -2,6 +2,7 @@ package xyz.hlafaille.seizuretracker.service;
 
 import jakarta.servlet.http.Cookie;
 import xyz.hlafaille.seizuretracker.entity.User;
+import xyz.hlafaille.seizuretracker.exception.UserEntityMissingException;
 
 import java.util.UUID;
 
@@ -9,8 +10,15 @@ import java.util.UUID;
  * Interface for interacting with Users
  */
 public interface UserService {
-    User getUserById(UUID userId);
+    User getUserEntityById(UUID userId) throws UserEntityMissingException;
+
+    User getUserEntityByEmail(String email) throws UserEntityMissingException;
 
     UUID createUser(String firstName, String lastName, String email, String password);
 
+    String encryptPassword(String password);
+
+    boolean isPasswordMatching(User user, String password);
+
+    void matchPassword(User user, String password);
 }
